refactor: finish changing proptype filenames
